requirements.txt
bitstring==3.1.5
dpkt==1.9.1
pypcap==1.2.2
PyQt5==5.11.3
netaddr==0.7.19